LanzAudit-Docker/app/requirements.txt
alembic==1.15.2
annotated-types==0.7.0
ansicolors==1.1.8
anyio==4.9.0
blinker==1.9.0
Brotli==1.1.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
cohere==5.15.0
cssselect2==0.8.0
fastavro==1.11.1
filelock==3.18.0
Flask==3.1.0
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
fonttools==4.58.0
fsspec==2025.5.1
graphviz==0.20.3
greenlet==3.1.1
gunicorn==23.0.0
h11==0.16.0
hf-xet==1.1.2
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.32.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.9
MarkupSafe==3.0.2
packaging==25.0
pillow==11.1.0
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
pydyf==0.11.0
PyMySQL==1.1.1
pyphen==0.17.2
python-dotenv==1.1.0
python-nmap==0.7.1
PyYAML==6.0.2
requests==2.32.3
sniffio==1.3.1
SQLAlchemy==2.0.40
tinycss2==1.4.0
tinyhtml5==2.0.0
tokenizers==0.21.1
tqdm==4.67.1
types-requests==2.32.0.20250515
typing-inspection==0.4.1
typing_extensions==4.13.0
urllib3==2.4.0
weasyprint==65.1
webencodings==0.5.1
Werkzeug==3.1.3
wheel==0.45.1
zopfli==0.2.3.post1
